重构IFieldCodec，collection支持codec

+++ nop-record/src/main/java/io/nop/record/reader/ITextDataReader.java
@@ -45,7 +45,7 @@ public interface ITextDataReader extends IDataReaderBase {
      */
     void reset() throws IOException;
 
-    default ITextDataReader subInput(int maxLength) {
+    default ITextDataReader subInput(long maxLength) {
         return new SubTextDataReader(this, maxLength);
     }
